added testscript for testfield
Built a mocha test-script for the textfieldView. Created a general-test
and test for the instances of the object.
There is a test for building instances with default parameters set, with
all parameters set, with some parameters set and with no parameters set. In these tests there is
also checked if the input types text, search and
date are used correctly. There is also a test for each
addToTemplate-function.